Python sir hohn (#1)
* some data to compare the model with

* further data to compare ICU capacities

* SIRD model for Germany

calculates the spread of the new corona virus
among the entire German population of 81.9 Million people.

Simplified assumption:
all infected cases are detected and no cases are undetected.

* running model version set up for Bremen. Using Bremen data.

* a model version for the Hubei data, which I used to constrain the recovery rate.

* additional tests to check the impact of undetected infections on scenario calculations

* model run on chinese data from the province Hubei

* model run for Bremen data. this run is to document
that the model is representing the development as
detected by the official data.

* model prognosis assuming that the spread of the
virus will continue as before restrictions came
into play. Contact rate is set to the same value
as at the beginning of the simulation as of day 50.

* model scenario assuming that all restrictions
currently active will be kept for the year to
come. The development in an isolated community
representing Bremen and direct surrounding is
shown here.

* This is a theoretical scenario to show that with
slightly released restrictions on contact rates
the number of infected people can be kept at a
constant  value and with this the model could be
tuned to match the intensive care unit capacity.
This simulation could be used as a guidance to
compare with the daily updated data to know if
restrictions can be released further or if
restrictions have to be tightened again.

But I am still working to match the intensive
care requirements with the model because the
data reported from the RKI do not match with
the current situation in Bremen.

* unstage files and move into new project folder

* data to compare the model with

* data on intensive care units in Bremen

* forgotten dataset from previous commit

* model code to run simulations

* some model results

* out commented irrelevant imports

* figures of hospitalized cases old (with data
from RKI on the relative fractions of
hospitalized and ventilated cases) and new
(adjusted to 22% hospitalized and 12% ventilated
according to the current situation in Bremen)

* documentation for github

* changed paths to figures

* changed equation in text

* another figure for the run for Germany with the official data

* adjusted link to correct figure

* corrected spelling mistake

* Update readme.md

Co-authored-by: Sönke Hohn <soenke.hohn@zmt-bremen.de>
Co-authored-by: Tillmann Muehlpfordt <tillmann.muehlpfordt@kit.edu>